LayoutTests/inspector/audit/audit-report.html

<!doctype html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/audit-test-fixtures.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("AuditReport");

    suite.addTestCase({
        name: "Instantiation with test suite",
        description: "should instantiate correctly.",
        async test() {
            let testSuite = new testSuiteFixture1;
            InspectorTest.assert(testSuite instanceof WI.AuditTestSuite, "testSuite is AuditTestSuite.");
            let report = new WI.AuditReport(testSuite);
            InspectorTest.expectThat(report instanceof WI.AuditReport, "Instantiate AuditReport with AuditTestSuite.");
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test for the AuditManager Instantiation.</p>
</body>
</html>
